memo changes
